Fix three crash bugs in Rust port; expand harness to 10 replays
**Bug fixes:**

seg005.rs: `run_jump` overflow — `(Char.direction as usize) + 1` wraps to
usize::MAX when direction is dir_FF_left (-1i8), panicking in debug mode.
Fix: cast through i32 first: `(Char.direction as i8 as i32 + 1) as usize`.
This crashed 5/9 doc replays.

seg002.rs: `do_auto_moves` misaligned pointer — `custom->shad_drink_move` is
inside the 1-byte-packed `custom_options_type`, so its address may be odd.
Accessing `auto_move_type` (align=2) fields through that pointer panics in
Rust debug mode. Fix: use `ptr::read_unaligned` for both field accesses.

seg006.rs: `find_room_of_tile` negative room index — when `curr_row_coll_room`
contains -1 (sentinel for "no room"), `get_tile(-1, ...)` sets `curr_room=-1`.
If tile_row>=3, the guard `curr_room != 0` is true for -1, triggering
`roomlinks[-2]` (OOB in C, panic in Rust). Fix: change guards to `curr_room > 0`
and clamp invalid rooms to 0 (no room) when boundary navigation is attempted.

**Harness:**

Add 9 golden traces from `doc/replays-testcases/` (levels 2, 5, 11, 12; covers
shadow, falling, grabbing, trick moves). Update run_harness.sh to run all 10
replay/trace pairs. Ignore `foretable`/`backtable` globally (known rendering
divergence in seg008.rs, visual-only, does not affect gameplay).

+++ rust/src/seg005.rs
@@ -930,7 +930,7 @@ pub unsafe extern "C" fn run_jump() {
         let xpos = char_dx_forward(4);
         let mut col = get_tile_div_mod_m7(xpos);
         for tiles_forward in 0..2 {
-            col += *core::ptr::addr_of!(dir_front).cast::<i8>().add((Char.direction as usize) + 1) as i32;
+            col += *core::ptr::addr_of!(dir_front).cast::<i8>().add((Char.direction as i8 as i32 + 1) as usize) as i32;
             get_tile(Char.room as c_int, col, Char.curr_row as c_int);
             if curr_tile2 == tiles_tiles_2_spike as u8 || tile_is_floor(curr_tile2 as c_int) == 0 {
                 let mut pos_adjustment =
